DEV V1

+++ unibet_test_technique/src/src/main/java/com/kindredgroup/unibetlivetest/repository/SelectionRepository.java
@@ -6,7 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SelectionRepository extends JpaRepository<Selection, Integer> {
+public interface SelectionRepository extends JpaRepository<Selection, Long> {
 
     List<Selection> getSelectionByStateEquals(SelectionState state);
+
 }
